Adjust type of link_len to match type of di_size in underlying struct, and change type of len to size_t to match strlen(3) result.  Drop unneeded casts.
u_int64_t is overkill because the actual size is still limited to SYMLINK_MAX
but it is probably better to match the filesystem types.

ok miod@, guenther@

+++ sys/lib/libsa/ufs2.c
@@ -455,8 +455,8 @@ ufs2_open(char *path, struct open_file *f)
 		 * Check for symbolic link.
 		 */
 		if ((fp->f_di.di_mode & IFMT) == IFLNK) {
-			int link_len = fp->f_di.di_size;
-			int len;
+			u_int64_t link_len = fp->f_di.di_size;
+			size_t len;
 
 			len = strlen(cp);
 
@@ -469,8 +469,7 @@ ufs2_open(char *path, struct open_file *f)
 			bcopy(cp, &namebuf[link_len], len + 1);
 
 			if (link_len < fs->fs_maxsymlinklen) {
-				bcopy(fp->f_di.di_shortlink, namebuf,
-				    (unsigned) link_len);
+				bcopy(fp->f_di.di_shortlink, namebuf, link_len);
 			} else {
 				/*
 				 * Read file for symbolic link
@@ -492,7 +491,7 @@ ufs2_open(char *path, struct open_file *f)
 				if (rc)
 					goto out;
 
-				bcopy(buf, namebuf, (unsigned)link_len);
+				bcopy(buf, namebuf, link_len);
 			}
 
 			/*
